+++ App.js
@@ -2,6 +2,8 @@ import React from 'react'
 
 import Home from 'components/Home'
 import Main from './components/Main'
+import About from 'components/About'
+import Contact from 'components/Contact'
 
 import { NavigationContainer } from '@react-navigation/native'
 import { createDrawerNavigator } from '@react-navigation/drawer'
@@ -15,7 +17,9 @@ function App() {
     <NavigationContainer>
       <Drawer.Navigator initialRouteName='Home'>
         <Drawer.Screen name='Home' component={Home} />
-        <Drawer.Screen name='Main' component={Main} />
+        <Drawer.Screen name='About Us' component={About} />
+        <Drawer.Screen name='Menu' component={Main} />
+        <Drawer.Screen name='Contact Us' component={Contact} />
       </Drawer.Navigator>
     </NavigationContainer>
   )