+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## ADDITIONS
 
-- Side Drawer
-- Home component
+- Contact Us and About Us components
 
 ## CHANGES
+
+- Drawer to account for the new comp.